feature: Add validation on HeaterMode

+++ routes/heaterMode.js
@@ -3,10 +3,10 @@
 var models = require('../models');
 var express = require('express');
 var router = express.Router();
+var util = require('util');
 
 
-//Get all HeaterMode
-router.get('/', function (req, res, next) {
+function findAllHeaderMode(req, res, next) {
   models.HeaterMode.findAll()
     .then(function (heaterModes) {
       res.send(heaterModes);
@@ -14,10 +14,26 @@ router.get('/', function (req, res, next) {
     .catch(function (error) {
       next(error);
     });
-});
+}
 
-//Create a HeaterMode
-router.post('/', function (req, res, next) {
+function createHeaderMode(req, res, next) {
+
+  // VALIDATION
+  req.checkBody('label', 'Invalid label').notEmpty().isAlphanumeric();
+  req.checkBody('order', 'Invalid order').notEmpty().isInt();
+  req.checkBody('max', 'Invalid max').notEmpty().isInt();
+
+  // SANITIZATION
+  req.sanitizeBody('label').toBoolean();
+  req.sanitizeBody('order').toBoolean();
+  req.sanitizeBody('max').toBoolean();
+
+  var errors = req.validationErrors();
+  if (errors) {
+    res.status(400).send({message: 'There have been validation errors: ', errors: errors});
+    return;
+  }
+
   models.HeaterMode.create({label: req.body.label, order: req.body.order, max: req.body.max})
     .then(function (heaterMode) {
       res.status(201).send(heaterMode)
@@ -25,10 +41,21 @@ router.post('/', function (req, res, next) {
     .catch(function (error) {
       next(error);
     });
-});
+}
 
-//Get a single HeaterMode
-router.get('/:heaterMode_id', function (req, res, next) {
+function findHeaterModeById(req, res, next) {
+  // VALIDATION
+  req.checkParams('heaterMode_id', 'Invalid id').notEmpty().isInt();
+
+  // SANITIZATION
+  req.sanitizeParams('heaterMode_id').toBoolean();
+
+  var errors = req.validationErrors();
+  if (errors) {
+    res.status(400).send({message: 'There have been validation errors: ', errors: errors});
+    return;
+  }
+
   models.HeaterMode.find({
     where: {
       id: req.params.heaterMode_id
@@ -44,11 +71,28 @@ router.get('/:heaterMode_id', function (req, res, next) {
     .catch(function (error) {
       next(error);
     });
-});
+}
+
+function updateHeaterMode(req, res, next) {
+  // VALIDATION
+  req.checkParams('heaterMode_id', 'Invalid id').notEmpty().isInt();
+  req.checkBody('label', 'Invalid label').notEmpty().isAlphanumeric();
+  req.checkBody('order', 'Invalid order').notEmpty().isInt();
+  req.checkBody('max', 'Invalid max').notEmpty().isInt();
+
+  // SANITIZATION
+  req.sanitizeParams('id').toBoolean();
+  req.sanitizeBody('label').toBoolean();
+  req.sanitizeBody('order').toBoolean();
+  req.sanitizeBody('max').toBoolean();
+
+  var errors = req.validationErrors();
+  if (errors) {
+    res.status(400).send({message: 'There have been validation errors: ', errors: errors});
+    return;
+  }
 
 
-//Update a HeaterMode
-router.put('/:heaterMode_id', function (req, res, next) {
   models.HeaterMode.find({
     where: {
       id: req.params.heaterMode_id
@@ -70,10 +114,21 @@ router.put('/:heaterMode_id', function (req, res, next) {
     .catch(function (error) {
       next(error);
     });
-});
+}
 
-//Delete a HeaterMode
-router.delete('/:heaterMode_id', function (req, res, next) {
+function deleteHeaterMode(req, res, next) {
+  // VALIDATION
+  req.checkParams('heaterMode_id', 'Invalid id').notEmpty().isInt();
+
+  // SANITIZATION
+  req.sanitizeParams('heaterMode_id').toBoolean();
+
+  var errors = req.validationErrors();
+  if (errors) {
+    res.status(400).send({message: 'There have been validation errors: ', errors: errors});
+    return;
+  }
+
   models.HeaterMode.destroy({
     where: {
       id: req.params.heaterMode_id
@@ -90,7 +145,21 @@ router.delete('/:heaterMode_id', function (req, res, next) {
     .catch(function (error) {
       next(error);
     });
-});
+}
 
+//Get all HeaterMode
+router.get('/', findAllHeaderMode);
+
+//Create a HeaterMode
+router.post('/', createHeaderMode);
+
+//Get a single HeaterMode
+router.get('/:heaterMode_id', findHeaterModeById);
+
+//Update a HeaterMode
+router.put('/:heaterMode_id', updateHeaterMode);
+
+//Delete a HeaterMode
+router.delete('/:heaterMode_id', deleteHeaterMode);
 
 module.exports = router;
